Adds Windows support
Before this change, we already had windows binaries available and some
functionality working. This makes the last adjustments needed for
Windows to pass both unit and end-to-end tests. While I can currently
use `make` personally, we shouldn't rely on it in Windows because paths
often have a space in them which makes make die. For that reason, the
GitHub Action is setup to use normal go instead.

Fixes #105

Signed-off-by: Adrian Cole <adrian@tetrate.io>